feat: crud login check

+++ src/main/java/com/ll/sb20231117/domain/article/article/controller/ArticleController.java
@@ -45,6 +45,8 @@ public class ArticleController {
 
     @GetMapping("/article/write")
     String showWrite() {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         return "article/article/write";
     }
 
@@ -58,13 +60,17 @@ public class ArticleController {
 
     @PostMapping("/article/write")
     String write(@Valid WriteForm writeForm){
-        Article article = articleService.write(writeForm.title, writeForm.body);
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
+        Article article = articleService.write(rq.getMember(), writeForm.title, writeForm.body);
 
         return rq.redirect("/article/list", "%d번 게시물 생성되었습니다.".formatted(article.getId()));
     }
 
     @GetMapping("/article/modify/{id}")
     String showModify(Model model, @PathVariable long id) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         Article article = articleService.findById(id).get();
 
         model.addAttribute("article", article);
@@ -82,6 +88,8 @@ public class ArticleController {
 
     @PostMapping("/article/modify/{id}")
     String modify(@PathVariable long id, @Valid ModifyForm modifyForm){
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         articleService.modify(id, modifyForm.title, modifyForm.body);
 
         return rq.redirect("/article/list", "%d번 게시물 수정되었습니다.".formatted(id));
@@ -89,6 +97,8 @@ public class ArticleController {
 
     @GetMapping("/article/delete/{id}")
     String delete(@PathVariable long id) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         articleService.delete(id);
 
         return rq.redirect("/article/list", "%d번 게시물 삭제되었습니다.".formatted(id));
